gestion des fichiers utilisateur

+++ backend/controllers/books.js
@@ -1,9 +1,15 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 // Créer un livre
 exports.createBook = (req, res, next) => {
+  const bookObject = JSON.parse(req.body.book);
+  delete bookObject._id;
+  delete bookObject._userId;
   const book = new Book({
-    ...req.body
+    ...bookObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   });
   book.save()
     .then(() => res.status(201).json({ message: 'Livre enregistré !' }))
@@ -26,18 +32,50 @@ exports.getOneBook = (req, res, next) => {
 
 // Modifier un livre
 exports.modifyBook = (req, res, next) => {
-  const book = new Book({
-    _id: req.params.id,
-    ...req.body
-  });
-  Book.updateOne({ _id: req.params.id }, book)
-    .then(() => res.status(200).json({ message: 'Livre modifié !' }))
-    .catch(error => res.status(400).json({ error }));
+  const bookObject = req.file ? 
+    {
+      ...JSON.parse(req.body.book),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+  
+  delete bookObject._userId;
+
+  Book.findOne({ _id: req.params.id })
+    .then(book => {
+      if (book.userId != req.auth.userId) {
+        return res.status(401).json({ message: 'Not authorized' });
+      }
+      if (req.file && book.imageUrl) {
+        const filename = book.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {});
+      }
+      Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+        .then(() => res.status(200).json({ message: 'Livre modifié !' }))
+        .catch(error => res.status(400).json({ error }));
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 // Supprimer un livre
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Livre supprimé !' }))
-    .catch(error => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then(book => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        if (book.imageUrl) {
+          const filename = book.imageUrl.split('/images/')[1];
+          fs.unlink(`images/${filename}`, () => {
+            Book.deleteOne({ _id: req.params.id })
+              .then(() => res.status(200).json({ message: 'Livre supprimé !' }))
+              .catch(error => res.status(400).json({ error }));
+          });
+        } else {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Livre supprimé !' }))
+            .catch(error => res.status(400).json({ error }));
+        }
+      }
+    })
+    .catch(error => res.status(500).json({ error }));
 };
